[core] update how the public api is organized
