+++ crates/mist-core/src/lib.rs
@@ -4,6 +4,3 @@ pub mod config;
 pub mod dialogs;
 pub mod parse;
 pub mod timer;
-pub use timer::timing;
-pub use timer::instant::MistInstant;
-pub use timer::run::Run;
