+++ crates/mist-core/src/parse/msf.rs
@@ -1,4 +1,4 @@
-use crate::Run;
+use crate::timer::Run;
 use ron::de::from_str;
 use ron::ser::{to_writer_pretty, PrettyConfig};
 use serde::Deserialize;
